Fixed up warning in project and changed from dll to static linking

+++ Anomaly Engine/src/Anomaly/imgui/ImGuiLayer.cpp
@@ -11,6 +11,7 @@
 //TEMPORARY=============
 #include <GLFW/glfw3.h>
 #include <glad/glad.h>
+
 //======================
 
 namespace Anomaly
@@ -29,11 +30,12 @@ namespace Anomaly
 		// Setup Dear ImGui context
 		IMGUI_CHECKVERSION();
 		ImGui::CreateContext();
-		ImGuiIO& io = ImGui::GetIO(); (void)io;
-		io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;       // Enable Keyboard Controls
+		ImGuiIO& io = ImGui::GetIO();
+		(void)io;
+		io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard; // Enable Keyboard Controls
 		//io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;      // Enable Gamepad Controls
-		io.ConfigFlags |= ImGuiConfigFlags_DockingEnable;           // Enable Docking
-		io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable;         // Enable Multi-Viewport / Platform Windows
+		io.ConfigFlags |= ImGuiConfigFlags_DockingEnable; // Enable Docking
+		io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable; // Enable Multi-Viewport / Platform Windows
 		//io.ConfigViewportsNoAutoMerge = true;
 		//io.ConfigViewportsNoTaskBarIcon = true;
 
@@ -45,13 +47,13 @@ namespace Anomaly
 		ImGuiStyle& style = ImGui::GetStyle();
 		if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable)
 		{
-		    style.WindowRounding = 0.0f;
-		    style.Colors[ImGuiCol_WindowBg].w = 1.0f;
+			style.WindowRounding = 0.0f;
+			style.Colors[ImGuiCol_WindowBg].w = 1.0f;
 		}
 
 		Application& app = Application::Get();
 		GLFWwindow* window = static_cast<GLFWwindow*>(app.GetWindow().GetNativeWindow());
-		
+
 		// Setup Platform/Renderer bindings
 		ImGui_ImplGlfw_InitForOpenGL(window, true);
 		ImGui_ImplOpenGL3_Init("#version 410");
@@ -62,7 +64,6 @@ namespace Anomaly
 		ImGui_ImplOpenGL3_Shutdown();
 		ImGui_ImplGlfw_Shutdown();
 		ImGui::DestroyContext();
-		
 	}
 
 	void ImGuiLayer::Begin()
@@ -77,18 +78,15 @@ namespace Anomaly
 	{
 		static bool show = true;
 		ImGui::ShowDemoWindow(&show);
-
-		
-		
 	}
 
 	void ImGuiLayer::End()
-	{	
+	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
 
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
-		
+		io.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()), static_cast<float>(app.GetWindow().GetHeight()));
+
 		//Rendering
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
@@ -96,13 +94,11 @@ namespace Anomaly
 		if (io.ConfigFlags && ImGuiConfigFlags_ViewportsEnable)
 		{
 			GLFWwindow* backupcurrentcontext = glfwGetCurrentContext();
-			
+
 			ImGui::UpdatePlatformWindows();
 			ImGui::RenderPlatformWindowsDefault();
 
 			glfwMakeContextCurrent(backupcurrentcontext);
 		}
-		
 	}
-	
 }
